gitlab_source/requirements.txt
fastapi
uvicorn
python-gitlab
pyyaml